[nixio] Disable bijective read-write test
The bijective read-write test now fails due to a bug in the object order
when iterating through the nix file. It appears to be related to the
closing of a file, which was introduced in a previous commit. Before the
introduction of the 'close()' method, the nix file would remain open
after the write portion of the test and the same file would be open
twice during read. With the 'close()' method implemented, after writing,
the file is closed and reopened for reading which inexplicably changes
the order of DataArrays (converted to AnalogSignals) when iterating.

I should note that HDF5 order tracking is enabled in the underlying NIX
library so this shouldn't happen. The test workaround should be
considered temporary.