Change the behaviour of `SpikeTrain.__add__` and `SpikeTrain.__sub__`
according to the discussion in #369.

When adding and subtracting Quantity scalars, `t_stop` and `t_start` are also modified by the amount of the scalar.
When adding and subtracting Quantity arrays, `t_stop` and `t_start` are not modified unless some of the new spike times fall outside the range [t_start, t_stop], in which case the range is extended to include all spikes.

Attempting to add or subtract two SpikeTrains raises a TypeError, except for the case of subtracting two SpikeTrains of the same size, in which case a Quantity array is returned.