[NixIO] Check if existing objects are already linked
Sometimes an object may already be stored in the nix file but not at
every location it should be. Until now, if an object was found to be
already in the nix file, no operation would take place.

We need to check for the case where the object is stored in the file
but is not linked to the location we're currently looking at (e.g., a
DataArray might have to be linked to multiple groups).