Fixed bug with BlackRockIO not loading certain files because of spikes before t_start or after t_stop
I fixed a bug that caused BlackRockIO to fail loading certain files because of spikes otuside of the time range set by the nsX start and stop times by setting SpikeTrain t_start and t_stop to the lowest or highest
spike timestamp present if they are outside of that range.
Apart from that I changed t_start from 0 to first spike timestamp if no nsX file is available.