Added ChannelIndex annotations for BlackRockIO
I added the required stuff to raw_annotations in BaseRawIO and implemented those in BaseFromRawIO.
BlackRockIO can now pass the correct annotations to the ChannelIndexes. In case of signal_group_mode='group-by-same-units' every keyword contains a list with multiple values.
This should not change anything for any other IO using the RawIO API