Removed references to files that do not exist on gin.
These files were duplicate, so I did not upload them to gin but never
removed them from the files that should be downloaded by the tests. I
never noticed, because tests were not failing for me. On Python3, they
are skipped, but on Python2 you would never notice anything because of
how urlretrieve works. Instead of failing with Error 404, it would just
download the HTML page showing an image with 'Error 404' or something.
This makes the files available and thus causes the test not to be
skipped. If one tried accessing them, the test would fail, because the
file is an HTML file.