warn if buggy nTelegraphEnable detected
A suggestion by @samuelgarcia on https://github.com/NeuralEnsemble/python-neo/issues/544

This fix improves support and logging for ABF1 files which may have buggy headers. If bugginess is detected (nTelegraphEnable a value other than 0 or 1), nTelegraphEnable is ignored and data is not scaled by fTelegraphAdditGain and the user is warned of the bug.