[nixio] nix.Value removal: New metadata compatibility
NIX no longer uses Value class for managing the value of metadata
properties. Values are added and retrieved directly.

Empty property values are now supported, but require a type to be
specified. The IO uses string as a default.

NIX returns collections of property values as tuples. The IO returns
lists explicitly for consistency with older behaviour.