Adapted BrainwaresrcIO for array annotations
This was the only IO that failed tests after event labels were changed
to use the array_annotation infrastructure.
Labels are always wrapped into an array now, but BrainwaresrcIO expected
them to be scalar if there was only a single one. So I changed it to
take the 0th element from this array.
I also implemented a check that tests if there really is only a single
label and throws an error if not.