WIP fix for #556
The `__eq__` methods for both AnalogSignal and IrregularlySampledSignal assumed that the arguments would be of the corresponding type. This allowed, for example, an AnalogSignal to be compared to another AnalogSignal using `==`, but these types could not be compared to scalars (e.g., `sig == 0.1`).

This is a problem in NumPy 1.15.0, since NumPy's `assert_array_equal` now uses comparisons to scalars to check for special cases, such as the presence of infinity in the arrays.

This commit changes the `__eq__` methods so that, before comparing attributes such as time parameters, they first check whether the type for comparison is a compatible signal type. Comparing times is skipped if the type is something else, such as a scalar.

This only partially fixes issue #556 and instead unmasks other problems. With these changes, the equality tests added in the previous commit pass for NumPy < 1.15.0 but continue to fail for NumPy == 1.15.0 for different reasons (namely, `assert_array_equal` wants to index into the signals with boolean arrays, which Neo does not support yet).