Fixed detection of warnings in test_blackrockio.py
Now no position is assumed any more instead the test checks if the warning
is in the list of warnings.
Apart from that the warnings.simplefilter filter should only be changed within
the context of catch_warnings here in order to ensure it being set to its original
state afterwards.