[ArrayAnnotations] Allow empty array_annotations
Before if a sequence of length 0 was passed, array_annotations were
created to match the desired dtype and the length of the object they
belonged to. Now instead of doing this and allowing for
misinterpretation of the values that were generated, an array of length 0
is array_annotated to the object.