Improved checking of lists as array annotations
Only the first element need to be checked if converting to numpy array
first. Cathing the case where the dtypes are so different that the
result dtype is object. After this only the first element of the array
needs to be checked for correctness, because the others share the same
dtype.
This increases performance.