Some improvements to AsciiSignalIO
in preparation for working on #579. Main changes:

(1) moved most arguments from `read_segment()` method to the IO class constructor, and added a `read_block()` method. This makes AsciiSignalIO more consistent with other Neo IOs.
(2) added an argument "multichannel", to control whether to create a multi-channel AnalogSignal or a list of single-channel signals. The default is the latter, for backwards compatibility.
(3) more consistency checks on argument values
(4) fixed a bug with "genfromtxt" where usecols would be applied twice.