No copy of parents when deepcopying data object
When deepcopying a data object, parents should not be copied, because then also all other children would be copied and also it is not expected behavior.
But keeping a reference to the original parent creates a unidirectional link which is not wanted either, so the references are just set to None.
The deepcopy behavior could maybe be unified in the DataObject class