[nixio] Discard 'neo_name' property on read
Small modification of the naming scheme described in PR #331:
In step 4, upon reading the 'neo_name' property and restoring the
original Neo object's name, the 'neo_name' attribute is now discarded.
Since we already restore this to the name of the object, there's no need
to keep it around in the object's annotations.  More importantly, it
would create a conflict on rewrite of an object, since the creation of
the property is handled explicitly, outside of the general annotation
conversion function.

Fixes #694