[nixio-fromraw] Trim nix_name annotation for chx objects
- The basefromrawio class creates ChannelIndex objects and appends
annotation values to existing annotations, one for each signal in the
channel.  This creates an issue with the 'nix_name' annotation which is
meant to be used as the NIX object name when the Neo structure is
written to a NIX file.  We override the 'read_block()' method from the
'BaseFromRaw' class to first call the original method and then trim the
'nix_name' annotation to make it a single value rather than a list.
- Re-enabled the 'test_write_proxyobjects' test since it is working now.