Fixing an integer overflow bug in `spike2rawio.py`
Due to the fact that the header data is read using numpy using numpys
fixed size integer types, any computation on them without a cast results
in computations on this fixed size types, which can result in over- or
underflows.

This can cause a bug in the `get_sample_interval` function of
`spike2rawio.py`, where multiple `int32` numbers are multiplied, which,
if they are large enough, can result in an overflow.

This commit fixes this bug by casting the first operant to `int`, so the
multiplication is performed on pythons arbitrary size integers.

As this function returns a float, another possible solution would be to
cast this into a float type, or, as it is later divided by 10^6, to move
that division to the front of the equation. This could potentially
increase the performance.