Update intanrawio.py
I added a `notch_filter_applied` info box as a boolean to indicate whether the Intan hardware applied the notch filter. 

For both rhs and rhd I just added the check at the bottom  before the return since we need to check both whether the `notch_filter_mode` > 0 and that the version is >= 3.0. I couldn't think of a way to fit it during the notch filter check without adding conditions into other functions which would slow everything down. It just means that the `notch_filter_applied` would be out of order compared to the other keys.